+++ Projet_Manic_Shooter/Assets/Scripts/Vagues/Vague0_Facile.cs
@@ -1,0 +1,21 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class Vague0_Facile : Vague
+{
+    // Start is called before the first frame update
+    protected override void Start()
+    {
+        base.Start();
+    }
+
+    // Update is called once per frame
+    protected void Update()
+    {
+        for (int i = 0; i < script_SousVagues.Length; i++)
+        {
+            
+        }
+    }
+}